pyproject.toml
[tool.poetry]
name = "fast-library"
version = "0.1.0"
description = "Fast and light library API"
authors = ["Robert Pacak <r.pacak@seznam.cz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
email-validator = "^2.2.0"
httpx = "^0.28.1"
pytest = "^8.4.1"
black = "^25.1.0"
mypy = "^1.17.1"
types-requests = "^2.32.4.20250611"
types-psycopg2 = "^2.9.21.20250718"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
